Feature/kernel implementation

+++ op_host/softmax_cross_entropy_with_logits_infershape.cpp
@@ -31,11 +31,11 @@ static ge::graphStatus InferShapeSoftmaxCrossEntropyWithLogits(gert::InferShapeC
     const gert::Shape* labelsShape = context->GetInputShape(IDX_1);
     OP_CHECK_NULL_WITH_CONTEXT(context, labelsShape);
 
-    int64_t dimNum = featuresShape->GetDimNum();
+    size_t dimNum = featuresShape->GetDimNum();
     OP_CHECK_IF(dimNum < 2, OP_LOGE(context, "features rank must be at least 2"), return GRAPH_FAILED);
     OP_CHECK_IF(labelsShape->GetDimNum() != dimNum,
                 OP_LOGE(context, "features and labels must have the same rank"), return GRAPH_FAILED);
-    for (int64_t i = 0; i < dimNum; ++i) {
+    for (size_t i = 0; i < dimNum; ++i) {
         int64_t featuresDim = featuresShape->GetDim(i);
         int64_t labelsDim = labelsShape->GetDim(i);
         OP_CHECK_IF(featuresDim >= 0 && labelsDim >= 0 && featuresDim != labelsDim,
@@ -49,7 +49,7 @@ static ge::graphStatus InferShapeSoftmaxCrossEntropyWithLogits(gert::InferShapeC
     OP_CHECK_NULL_WITH_CONTEXT(context, backpropShape);
 
     lossShape->SetDimNum(dimNum - 1);
-    for (int64_t i = 0; i < dimNum - 1; i++) {
+    for (size_t i = 0; i < dimNum - 1; i++) {
         lossShape->SetDim(i, featuresShape->GetDim(i));
     }
